Refactor OffscreenPuppetService.  It was getting bloated and now there is a needed for a separate MacOS renderer.
